backend-AI/web/requirements.txt
flask
requests
beautifulsoup4
trafilatura
psycopg2-binary
qdrant-client
python-dotenv
